Gráfico Misto
Novo projeto para o gráfico misto, novo gráfico com número de amostras exponencial
