+++ graficoMisto/graficoMisto.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32505.173
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "todosSorts", "todosSorts.pyproj", "{533DE687-16D4-4B2A-AAF4-CCBF2B32DADA}"
+Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "graficoMisto", "graficoMisto.pyproj", "{73EAF540-D657-4913-B724-F3089DE0AB89}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,13 +11,13 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{533DE687-16D4-4B2A-AAF4-CCBF2B32DADA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{533DE687-16D4-4B2A-AAF4-CCBF2B32DADA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{73EAF540-D657-4913-B724-F3089DE0AB89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{73EAF540-D657-4913-B724-F3089DE0AB89}.Release|Any CPU.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D0685D05-3F19-42C5-8714-2DCD25FA6F88}
+		SolutionGuid = {363E8E3D-03F5-424A-AFCD-E89DF510104E}
 	EndGlobalSection
 EndGlobal
