+++ graficoMistoExponencial/graficoMistoExponencial.sln
@@ -1,0 +1,23 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32505.173
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "graficoMistoExponencial", "graficoMistoExponencial.pyproj", "{E9C076E9-9D99-47C4-A56B-297178EA806E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E9C076E9-9D99-47C4-A56B-297178EA806E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E9C076E9-9D99-47C4-A56B-297178EA806E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {515849EA-BE60-463C-8AFF-B11756E19E2D}
+	EndGlobalSection
+EndGlobal
