--- todosSorts/saida.txt
@@ -1,1 +1,0 @@
-[[10, 110, 210], [0.0, 0.0078125, 0.0234375]]
